pyproject.toml
[tool.poetry]
name = "really_estate"
version = "0.1.0"
description = ""
authors = ["Chase <robertchasesommer@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^1.5.1"
requests = "^2.28.1"
beautifulsoup4 = "^4.11.1"
ipykernel = "^6.17.1"
jupyter = "^1.0.0"
lxml = "^4.9.1"
prefect = {extras = ["aws"], version = "^2.7.0"}
prefect-aws = "^0.2.0"
black = "^22.10.0"
modal-client = "^0.44.600"
google-cloud-storage = "^2.7.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
